hide model behind interface
